The rewrite of SysWinGetPrinters() lost a little in the translation.  A valid condition is tagged as an error.  This restores the 4.0.0 behavior, which does work.
Note, this code look wrong, although it works.  I think it is some Windows
95 trick.  I intend to rewrite that section tomorrow.  1.) W2k and on
should be using printer level 4, not 2.  2.) I would think the correct way
to code this would be to call EnumPrinters() with a buffer size of 0 to
get back the size needed.  Then malloc that amount and call EnumPrinters()
again.