SysFileCopy for Unix : preserve timestamps, like Windows version <==> cp --preserve=timestamps fromFile toFile. symbolic links are always dereferenced (not sure it's like Windows version - see MSDN CopyFile).
SysFileMove for Unix : 
if fromFile and toFile are on the same filesystem then do rename, otherwise do copy+unlink because rename does not work across filesystems.
more work probably needed in case of symbolic links moved across filesystems : the copy+unlink sounds not natural because the symbolic links are dereferenced.