Merge revision 5974 from trunk into 4.1.0:
This commit changes (slightly) the way the search for the base image is
done on Windows.  It fixes a problem I am seeing when I build with
the image either not being found, or the wrong one being loaded.

-    if (!SysFileSystem::primitiveSearchName(BASEIMAGE, getenv("PATH"), NULL, fullname))
+
+    if (!SysFileSystem::primitiveSearchName(BASEIMAGE, NULL, NULL, fullname))

With the old code, if there is a rexx.img in the path, it is picked up
before the image in my build directory.

The change results in the Windows API SearchPath() being called in a way
that looks first in the directory that the application is loaded from (which
is rexx.exe.) If not found there, it still searches the path.

Since on Windows, rexx.img is always in the same directory as rexx.exe, the
change should, if anything, result in the image being found sooner.

This is only in SystemInterpreter::loadImage() so it has no effect on the
search for anything other than BASEIMAGE.