This update is the final solution to get ooRexx libraries installed into /usr/lib (or /usr/lib64) instead of /usr/lib/ooRexx. The solution was a two stage process.
We added one statement to configure.ac which forces a substitution on pkglibdir and changes it to libdir. The substitution happens during the configure step so it always overrides the default pkglibdir setting.

Then we changed the rpm spec file, the deb files, and the Mac files to remove all references to /usr/lib/ooRexx and change them to just /usr/lib.

NOTE!!!!!!

This has only been tested on RPM distributions. I will make any updates later to the deb files after tonights build machine builds. The Mac files will need to be tested by hand.