3529261 Win32 V4-1-1 uninstaller issues
Give more space to the label to acount for very long path names.  But,
if the user installs ooRexx into a very very deep directory the label,
the warning text will be clipped.  There is nothing to do about that,
the installer only has so much space on a page.  Normally people do not
install into a directory nearly as deep as the user that opened the bug.