Add a new ooRexx extension: ooSQLite
ooSQLite provides a complete interface to SQLite.

SQLite is a self-contained, serverless, zero-configuration, transactional
SQL database engine.

This database is embedded in the ooSQLite package, which means that any
ooRexx program that requires the package by using:

::requires 'ooSQLite.cls'

has everything it needs to create and use databases through a highly
functional database engine.  No addtional libraries or software needs to be
installed on any platform to use ooSQlite.  The SQLite database engine has a
very small footprint, making the implementation library for ooSQLite well
under 1MB in size.

ooSQLite will be available for all platforms that the interpreter is
supported on.  The package provides both an object-orientated interface and
a 'classic Rexx' interface.  Technically, the 'classic Rexx' inteface should
be called a 'functional' inteface because it uses ::routine() and in a very
few places some array and directory objects.

At this time the package is functional and useful.  The SQLite API has a few
core APIs and a large number of APIs that are only of use in very rare
circumstances.  The majority of all the APIs are accesible through ooSQLite
at this point.  The intent is to provide access to the complete set of APIs
at some point.

Mostly tested on Windows at this point, but the Linux version builds and has
been lightly tested.  The nature of the SQLite API and the ooSQLite
implementation library is such that there really should not be anything that
works on Windows that would not work on the other platforms, if the library
builds without error.