foodsPristine.rdbx                              ------------------
The binary database file, foodsPristine.rdbx, was created from a
database definition file, foods.sql, included in the example code that
accompanies the book:

"Definitive Guide to SQLite", Second Edition, by Grant Allen and Mike
Owens, published by APress.

APress and the authors have okayed the use and inlcusion of the binary
file in the ooRexx project.

Reference ticket: XWL-89783

Message to helpdesk@apress.com
------------------------------

I've bought the The Definitive Guide to SQLite, Second Edition, new from
Amazon.  As a reader I've downloaded the source code to accompany the
book.

I am also a committer on the Open Object Rexx project:
http://sourceforge.net/projects/oorexx/
http://www.oorexx.org/contact.html

I've written an ooRexx binding to SQLite.  For some of the example
ooRexx programs for the binding that I'm writing, I would like to use
the foods.db database file.  The database file itself was created using
the foods.sql file, which is included the source code that accompanies
the book.

Note that the foods.db file is not included in the source code, only
foods.sql is included.

ooRexx is an open source project.  For the distribution of the example
ooRexx programs, I would like to include the binary foods.db file with
the examples, along with an attribution to the book, the authors, and
Apress.

So, I'm asking your permission to do that.

It's not clear to me if Apress or the authors would object to that use
of the binary foods.db file to begin with, since it is not a part of the
source code any way.  I don't know if I need to contact the authors
separately, or contacting Apress is sufficient.

If it is okay to include the binary foods.db file, but you would rather
I don't give an attribution, of course that is fine with me.

I'd appreciate any information you can give me on this topic.

Thanks.

--
Mark Miesfeld
miesfeld@gmail.com


Reply #1 from helpdesk@apress.com
---------------------------------
Dear Mark Miesfeld ,

Thank you for contacting Apress.  This is an automated response
confirming the receipt of your ticket.

For your records, the details of the ticket are listed below.  If
replying, please make sure that the ticket ID is kept in the subject
line to ensure that your replies are tracked appropriately.

Ticket ID: XWL-89783
Title: The Definitive Guide to SQLite - question

With best regards,
Apress Customer Support

Reply #2 from Apress Customer Support support@apress.com
--------------------------------------------------------

Mark,

It's okay, we can look it up.

So you just want to use a database file?  That should be no problem,
particularly since you plan to properly attribute the authors and
Apress.

Please let us know if you have any further questions or concerns.

With best regards,
Apress Customer Support